requirements.txt
python-dotenv>=0.21.0
requests>=2.0.0
beautifulsoup4>=4.0.0